requirements.txt
# requirements.txt (Production)

# Flask Framework
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1

# MongoDB
mongoengine==0.27.0
pymongo==4.6.1

# Configuration
python-dotenv==1.0.0

# Security
Werkzeug==3.0.1
email-validator==2.1.0.post1

# Data & Export
pandas==2.1.4
reportlab==4.0.7
openpyxl==3.1.2

# API Communication
requests==2.31.0

# Machine Learning
scikit-learn==1.7.1
joblib==1.3.2
imbalanced-learn>=0.12.0  # For SMOTE (updated to latest compatible)

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dateutil==2.8.2

# SSL and Networking
certifi==2023.11.17
urllib3==2.1.0

google-generativeai>=0.3.0                  # Gemini integration
python-whois==0.8.0              # WHOIS lookups
OTXv2==1.5.12
gunicorn==21.2.0

redis==5.0.1
celery==5.3.4
hiredis==2.2.3
flower==2.0.1
quart==0.19.4
hypercorn==0.15.0